More accurate notification panel logs.
Trigger notification panel logging based on panel expansion and
non-dozing, rather than visibility and interactivity. Turns out this
is more specific to the cases where the panel is actually shown, in my
testing.

Turns on/off notification visibility logging if the panel is believed
to be on, but also respect existing StatusBar hints about visibility logging.

Added new test for not-lockscreen panel log.

Bug: 155108241
Test: atest NotificationLoggerTest
Change-Id: Ic418417afc6cac073429938885773c56aa55fef8

Former-commit-id: d93cb57302a8e71d8b819f33fe276f12982cd18d