Fix returning an occluded activity below boundary
WindowContainer#getActivity() was returning null if there was no
matching activity or hit the boundary. Since we were unable to
tell the difference between the two cases, we would continue looking
for a matching activity from higher level window containers.

Take below hierarchy for example, #getActivity() returns null if
the boundary was TopActivity. So, we were kept looking for the matching
activity from Task#2 and returns BottomActivity.

RootTask
  - Task#1
     - TopActivity
  - Task#2
     - BottomActivity

Bug: 156014697
Test: start activity from recents in split-secondary
Change-Id: Ic344e8c1c47e9fa05c4de60be8c6b30f75cdb0bc

Former-commit-id: 62cdfe75f611cb667e9a42598da958170b400e91