Use java_sdk_library for API for updatable-media
The API for the updatable-media library is prefixed with
framework-media so it cannot be switched to use java_sdk_library
directly. Instead a new API only java_sdk_library was added for
framework-media.

The names of the individual modules do not quite follow the pattern
that java_sdk_library uses so this temporarily sets the following:
    naming_scheme: "frameworks-modules"

That causes java_sdk_library to use a naming scheme that matches the
one used by the individual modules of this. It will be cleaned up
later.

Test: m update-api
Bug: 155164730
Change-Id: Ib9dd6d276856146e161578697cc6ba83b26e4ad8

Former-commit-id: 289a482fea6a4c0772db4b45529798be33b96980