Always open bubble for bubble notifs + fix issue
Clicking on the notification will only redirect to the
bubble if the notification has FLAG_BUBBLE. This changes
the logic so that it will happen if the notification is
allowed to bubble.

This also includes a fix for an issue where we would
remove the bubble (and notification) incorrectly on updates:
Previously we'd always remove a bubble if the notif was
updated & no longer had FLAG_BUBBLE. There are some
situations where a notif might not have the flag but also
shouldn't be removed (e.g. bubble dismissed). This CL alters
it so that the bubble is removed when canBubble is false
instead of FLAG_BUBBLE being removed.

Test: atest SystemUITests
Bug: 149033291
Bug: 152883583
Change-Id: If9c61295d09dfa9f36360b02bcd66967bcb7d667

Former-commit-id: 76343014ce9f66d5d343eac1828563962aa69dcc