Add keylayout for Steam controller (USB)
In the USB mode, Steam controller needs a key layout, because most of
the buttons and axes are incorrectly mapped.

The wireless dongle has not been tested.

The controller also support BLE, but currently goes into lizard mode by
default (only mouse and 2 buttons: left/right clicks work). So a driver
is likely necessary to go into the normal mode over BT, and hid-steam
doesn't help with that currently.

On BT, the controller has product id of 0x1106.

Bug: 147434575
Test: use the "gamepad tester" app and connect joystick over the USB.
Ensure that all mappings work
Test: CTS test will not be done at this time because it requires
supporting SET_REPORT in the hid command.

Change-Id: Ic0c50695338f98dcaeaaa7b3417fde702563d205

Former-commit-id: 2abc5d2823dc3cb158d7fe3525a494c93fd9c0ee