Camera: Fix race for onCaptureBufferLost callback
The callback holder was removed when the capture sequence is
completed, which is too soon because the buffer loss callback could
potentially arrives later than the capture sequence completion.

Defer the deletion of the callback holder to when the native inflight
request is removed, which takes into consideration of error
notifications.

Test: Camera CTS
Bug: 155353799
Change-Id: I12b716acc9fbaa9791f0498ac77d4470a7448d5a

Former-commit-id: db97350aa110c2fc888de04d9c3c98c1994e368f