Revert^2 Fix missing animation when launch activity from notification
The remote animaton is controlled by notification shade, if we set
the notification shade set to invisible, although the animator is
still running, but surface won't update.
Extending the visibility time until the animation finish can
solve this problem.

Also post the calling of setLaunchResult from main thread so
mIsLaunchForActivity and mAnimationPending can be synchronized.

Fixes: 154815513
Test: atest SystemUITests
Test: Follow the issue description.
Change-Id: I735340498575e2d7dc5a95ec1be3aa886739ecc4

Former-commit-id: a174e882511cc033a5de13d7b2a91fe68b78c8f5