Update home bounds when entering split on rotation
Sometimes the rotation call comes in before split
is fully entered. However, we already know that we
are going into split-mode, so use a partial split
check rather than the full check to determine
whether home should be updated.

Also fixes another race where the incorrect position
was used to find closest snap on rotate before that
snap was reset to middle when leaving split (due to
a previous change that sets mMinimized early).

Bug: 156491146
Test: make various attempts at entering split by
      opening apps in a different orientation.
Change-Id: I380653e47dc0d07488c3ac56d39c4571d2f25f0b

Former-commit-id: 90d5b6d0ca79f180a108a325065de33be38de07b