Reset default browser on installExistingForUser
Previously the preferred browser was only reset on the initial
install of a package. So if a different user already had the browser
installed, adding it to another user would not reset their default
browser.

In general the post install steps should be split up into
postFirstInstall and postInstallForUser, where the former calls the
latter, but that's a much larger change.

This doesn't fix an unfortunate bug where uninstalling a browser
and re-installing it without re-querying for the preferred browser
will clear the browser role, but not wipe the preferred Activity from
settings. Fixing this requires a much larger refactor, and is probably
not worth it given the assumption that users don't re-install apps
expecting it to reset the browser.

Bug: 152898545

Test: manual install browser to personal profile and then to work profile,
    verify disambiguation Activity pops up

Change-Id: Ie02b45c28c705bdcfb5bf89b9e7a33b4850cf0f7

Former-commit-id: 959c54f49112322b9caaac962a6a3bd2fabfd782