Bubble's change could notify to A11y framework
The virtual display of the activity became private and A11y framework
won't track the windows of the private virtual display. Then the change
of bubbles using the activityView could not be notified to A11y
framework.

Due to the bubbles created from system ui, we made the private virtual
display created from system ui could track its windows.

This is the short term solution, and only phases in rvc-dev branch. We
plan to track all private virtual displays as the long term solution at
master branch, but need to avoid this issue happened again, b/148760652.

Bug: 156685691
Test: a11y CTS & unit tests
Change-Id: I7a7e3d6040ce05f5d3769c14d34f906d8042c473

Former-commit-id: 167804a645ef62ea1dc4d0e30f43082c49bf6c2c