Add multi-user support for AutoTileManager
Before this CL, AutoTileManager (and AutoAddTracker) would only work
"corectly" for user 0, as all the settings (read and write) were done in
user 0. This meant that tiles that got auto added as user 0 would not be
auto-added for other users. Similarly, tiles that got auto added as a
secondary user would not be auto added as user 0.

This CL makes sure that everything is done with respect to the current
user (as reported by QSTileHost).

Test: manual
Test: atest AutoTileManagerTest AutoAddTrackerTest
Bug: 154607890
Change-Id: I518334cca0a0960e0b3b74b43ac9fce4610c7bd0

Former-commit-id: 1b5d7afb7f118a63b2fd7b634d0c6b04dbf0f67b