Avoid #finalResizePip after dismissal of PIP.
When dismissing PIP via the close button or via dragging to the dismiss
area, #finalResizePip is being called which then sets the PIP task back
to its PIP bounds when it should be set undefined already by the
framework.

Tap to close: The tap UP event causes a double-tap delay callback, which
gets ran shortly after the close button is tapped. Since we are already
dismissing and PipTouchHandler gets that callback, we will simply remove
the double-tap callback. We will also shorten the hideMenu animation to
0 for this case.
Drag to dismiss: The dismissal causes an animation, which then calls for
a Resize-callback. There really is no need to do this for the case of
dismissal, so we will avoid calling it.

Bug: 155794931
Test: Dismiss both PIP and then try to get PIP enter secondary
split-screen, and YouTube no longer stays black screen

Change-Id: Ic93b210f9f3f7d8947f3fe36f5fa32886f8170ee

Former-commit-id: cd3f2de8b88cebcf10532084d887bd6bf41430f0