Added getSyncTransaction
Create a getSyncTransaction that will return the blastSyncTransaction if
in a sync or the regular pendingTransaction if not. This is to separate
situations where we want to directly call getPendingTransactions so the
transaction is not synced with the client request.

One example is WM or remote animations starting. We don't want the
animation setup to be synced. If it syncs with the client request, the
initial animation setup will be delayed, but the animation will still
start running since it thinks everything has been set up properly. In
those cases, continue to use getPendingTransaction instead of
getSyncTransaction

Test: Animations run during applySyncTransaction
Bug: 153579514
Change-Id: Ied21f3d2addb19e7bb814379567de076d429007f

Former-commit-id: 89cc30e994b924124f7e8f725a35c337bf97e58e