Fix size compat bounds in close-to-square display
DisplayContent#ignoreRotationForApps does not mean the display
cannot rotate. It means that the display can rotate without
considering the orientation requested by application. Currently
the case is that the aspect ratio of display is close to square.

Originally if the close-to-square display is rotated, an activity
with fixed aspect ratio will still use non-rotated bounds that
causes unexpected scaling. This change makes the case calculate
the bounds by current rotation.

Fixes: 156543028
Test: SizeCompatTests#testFixedAspectRatioBoundsWithDecorInSquareDisplay

Change-Id: Id22522f9a142806a922f643f4802971df302dadb

Former-commit-id: f9fb7e7b2f994315de3cd976e45e319b4bbf7ade