Ensure screenshot edit action gets permissions for current user
Previously, edit action intent was sent to ActionProxyReceiver with the
SYSTEM user. It had the FLAG_GRANT_READ_URI_PERMISSION flag, but when
the current user wasn't the system user, the system user wouldn't have
permissions to read that URI so it didn't work. This caused Markup to
fail to open the screenshot since it wasn't given permission.

Bug: 148128165
Test: Validate that Markup can read/write/delete screenshot w/ secondary
      user.
Change-Id: Ifa1d8fd195d5f777ab51972183ed6db4e6735ffd

Former-commit-id: c57f0cb7f210fdc63d0515311c9fff41e5742b24