Don't save touch location as a resting position.
When you drag bubbles out of the magnetic target, it springs towards your touch position to 'catch up', since it was stuck in the target away from the touch. However, unlike all other springs, this one should not save its end position as the stack resting position, since it's not a valid resting position.

This one was hard to repro - you have to drag bubbles into the target, then quickly pull them out, back in, and release them before they catch up to your finger. If you do though, you'll end up with bubbles that come back in the middle of the screen.

Test: mash on the dismiss target a bunch
Bug: 155495386
Change-Id: I8ded5337a72fb52a851e7640d2b0fddb8453648a

Former-commit-id: c81cc482ddb72cdda0e8e7d47c1be6bd9f767f70