Paged carousel for media players
This is a stopgap implementation on top of HorizontalScrollView. It
will take much more work to move to a ViewPager2 implementation. In
the meantime, this at least adds a page indicator to that users will
realize that there are multiple pages.

Added in this change:
 - page indicator
 - fling to next page
 - centering current page when scrolled into

Bug: 153170704
Test: manual - Paged between three players in the carousel. Verify
that the page indicator showed correct page.
Test: manual - Scoll into player and release. Verify that page snaps
back to center.

Change-Id: Iae51d9136184a0a7de083177ea105f2db9a72c99

Former-commit-id: 52dfb62937831295761181131e593e9deaab16c8