Move pip behind home stack on dismiss
If home stack is nested, it means we're in split or some
other non-fullscreen mode. In these cases, dismissing a pip
task should move it to the back of the home-stack parent
rather than the back of the display.

This behavior is the same as what was done for Activity.moveToBack,
so this CL pulled that functionality into positionTaskBehindHome()
so it could be shared

Bug: 157159929
Test: ActivityStack tests. Or attempt repro in bug.
Change-Id: Icf04f87edf4c00b47e3f1e0a45d80bf9b680ce1c

Former-commit-id: 38257988f50c4375ad055fa4851d17abeeb18c82