Fix issue where media would not expire
We only register media listerners for active streams, and this can be a
problem. A stream might arrive not playing yet because it's buffering,
and we won't receive any other state notifications about it.

Let's register listeners for all media streams that have notifications,
regardless of them being already active or not.

Test: play/pause song without pulling down the shade, wait for media timeout
Test: play/pause song with shade down, wait for timeout
Bug: 153897770
Change-Id: I1809f2380bbe2840640baa732360a230904b0c57

Former-commit-id: 41c3f9c078e1d662cfb0eae6a61a33560e26e8eb