Notify that keyguard is dismissed when user is unlocked
Follow the steps of UsbDeviceManager and call
onKeyguardStateChange(false) from onUlockUser(...) in
StorageManagerService.
Without this, on the devices that do not have keyguard (such as TVs)
StorageManagerService.mSecureKeyguardShowing never changes, and remains
'true'forever (set initially to 'true' due to security concerns), also
the this incorrectly set value gets passed to vold, causing the latter
to skip newly added disks.

Bug: 156220570
Test: ran manually, attached USB drive, monitored vold logs
Change-Id: Ib9b04c5fcd173da9894084002400c691a5c10b57

Former-commit-id: 33e2b6ae44d428993e0e5c8c9b33d6266b7900d4