SnapshotMergeReported: add retrofit and cow image size
Additional information are important to evaluate the impact of Virtual
A/B on devices especially related to OTA take rate.
Take rate is strictly related to the space that is temporarily allocated
in the /data partition (cow_file_size_bytes), as well as knowing if the
device has been upgraded to use Virtual A/B (is_vab_retrofit) or if the
feature was natively supported.
Add these new fields to the SnapshotMergeReported metric.

Bug: 154016862
Bug: 157720191
Test: m
Signed-off-by: Alessio Balsini <balsini@google.com>
Change-Id: Iae8310a8f5364ee94fc8ad6e55362d86cb0752d8

Former-commit-id: 2294f39e2a27758d7b1c698594414096028d34ee