Add log to help debug bind TCS twice.
In b/156555486, we find two consecutive requests may cause bind TCS
twice. It is wired mBinding is false here, it should be true when
TCMS binds the TCS for 1st request. The only possibility is that
willBind is false or onServiceDisconnected, onBindingDied or
onNullBinding was called to set mService null and mBinding false.
The test fail cannot be reproduce locally with many tries. Add log
to help debug the same bug next time.

Bug: 156555486
Test: Hard to see the fail case, see local added normal log
Change-Id: I04600b00322ef2dcc2162770672c797d42f5e98c

Former-commit-id: 2e1870c23b57e158503d7624ffc2b6d62cc2db29