Check for null parent when migrating SurfaceControl
When migrating the SurfaceControl, the WC parent can be null. In this
case, we create a new SurfaceControl, but set it's parent to null so the
layer is still available, but will be offscreen. This way if the parent
becomes non null, the SC can get re-parented to the new parent.

Test: Split screen with removing secondary app
Bug: 153579514
Change-Id: Id0e2f2d5e5b7448c3cafd5b90120777f5175153c

Former-commit-id: 0a947d95901688be1b864247dea942b5dff2697c