Binder cache for isUserUnlocked update
Bug: 154434809

This change ensures that the isUserUnlocked binder cache will be
invalidated when a new UserManagerService or StorageManagerService is
created.  Note that the bug cannot be reproduced at will so it is not
known if this change actually fixes the bug (which has multiple
potential root causes).  However, by inspection, this change is
desirable.

Test: Run tests with an instrumented build.  No cache inconsistencies
were found:
  * UserManagerServiceCreateProfileTest
  * UserSystemPackageInstallerTest
  * UserManagerServiceUserInfoTest
  * UserManagerServiceIdRecyclingTest
  * PtsChreTestCases
  * UserLifecycleTests

Change-Id: Ibc1204ccc1458028d94f2a64cbdc8ec1234b1ef3

Former-commit-id: 5473e177368153ff172e248decc84599d3e57270