Add occlusion logic to OverlayViewGlobalStateController and CarKeyguardViewController
Activity has a public API setShowWhenLocked, which, if set to true,
allows it to be shown when the device is locked, thereby "occluding"
the Keyguard.

In the Phone OS implementation, when an occluding Activity is showing,
Keyguard is temporarily dismissed and StatusBar expands.

In the AAOS implementation where SystemUI components are mounted to
the SystemUIOverlayWindow, OverlayViewGlobalStateController hides all
views mounted to it unless they are configured to be shown even during
occlusion.

Test: Unit Tests + atest CtsWindowManagerDeviceTestCases:ActivityLifecycleTopResumedStateTests#testTopPositionLaunchedOnTopOfLockScreen
Bug: 156781505
Change-Id: I8320e97a575a990ba4301729c6b0e6c86d6ef7c5

Former-commit-id: 5dc67de14458ff5c6ffaab55f252432da036d082