Handling hearing aid devices correctly
This CL handles broacasts on the Hearing Aid profile, which was ignored.

In addition, we don't receive the broadcast about the bond state of
BT devices since what we're interested in is "connected devices" not
"paired devices".
When a new device is paired/forgotten, it becomes connected/disconnected so
it is sufficient to see connection state of devices.

Bug: 157282893
Test: manually w/ A2dp devices repeating pair/forget.
Change-Id: Ie66544ba4cb82fc6002f6265613df1e7f1201a92

Former-commit-id: e77fbcba188151b47224dca010e3c84745fcf359