Allow creating customized DA for Feature
(3/n DisplayArea Sibling Roots)

By creating customized DA for Feature, we can modify DA of different
Feature. Before, the DA.Root was dimmable, which caused the magnify
window not mirror the dim layer. Now, we make the DA of
FEATURE_WINDOWED_MAGNIFICATION dimmable, so that magnify window also
mirrors that.

Bug: 156425461
Test: atest WmTests:DisplayAreaPolicyBuilderTest
Test: verify the magnify window also mirrors dim layer.
Change-Id: Id0ec4acb2fe302bbcc2c222c5da94f99cb425324

Former-commit-id: 915c2ade8f30e53f1b807bab524f9735971c05cd