Create a binder cache for IBluetooth.getConnectionState
Bug: 157935587

This adds a binder cache for the IBluetooth.getConnectionState() method.
There is no change in bluetooth functionality.

Each atest is run twice.  Once with the code to be committed and once
with a special build that sets PropertyInvalidatedCache DEBUG and VERIFY
to true.  In the latter case, the test passes if the atest passes and if
there are no errors from the cache verification.

Tag: #feature

Test: atest BluetoothInstrumentationTests

Test results are
  Passed: 479, Failed: 0, Ignored: 4, Assumption Failed: 75

Change-Id: I2946297ffab557877dc7ec56206834d7c3776662

Former-commit-id: 99e9f58e24ca354ed5a42e3723b23e9a995faf1e