Remove IoUtils from service-permission system-server stubs
As this did not have @hide this was being treated as if it was part of
the public API and so while it was not being tracked in a .txt file
because only the system-server API was being generated it was being
included in the system-server stubs as metalava will include public API
stubs in the stubs for every API surface.

Adding @hide ensures that it is no longer exported as part of the
service-permission stubs.

Bug: 155164730
Test: m service-permission-stubs and verify that it does not include IoUtils
Change-Id: I1ffa284c1f692397b52bdd6b3750b13125887352

Former-commit-id: 1835fdc3dbee497e295ddc36e00af8cc091dfd8a