Allow blocking calls to BackupTransport as a temporary measure
To protect system stability, any Binder calls leaving the
system_server should be one way. The ideal solution
would require redesigning the IBackupTransport interface
which might be a major undertaking.

But meanwhile, we still need to continue making these
calls. So, as a temporary measure, this CL allows blocking
calls to be made.

The long term solution of redesigning the interface so as
to make the calls non-blocking can continue to be tracked
on the same bug.

Bug: 147702043
Test: atest TransportManagerTest

Change-Id: I5c7ec5b26d6ad06608576e4d98dd439992ae7dfe

Former-commit-id: ab10e44b3f6cebc3a25ce80880d399acc9a844dc