Call WindowContainer#dump in all levels
In order to know if there is an animating window container. And
all derived classes should call super because the dump method is
annotated as CallSuper.

This fixes that TaskDisplayArea and ActivityStack don't call super
and the overridden Task#dump is never called. Also adjust the output
format so it is easier to read.

Bug: 157743122
Test: "adb shell dumpsys activity activities" and check the section
      "Task display areas in top down Z order".
Change-Id: I8396e1fbce1960dc9dcd9de8c5cd47bfd18ba31a

Former-commit-id: 4f088b6c3083a0b11c13e665cad22c99ac6ba8f7