Ignore rotation update during recent animation.
During the recent animation, the closing app is still considered on top
and its required orientation is used for the next rotation computation.

When going to launcher and rotating the device at the same time, a
rotation update is triggered by the sensor and if the closing app allows
a change of orientation, the display ends up rotating at the end of the
animation, which make the launcher rotate even if it is not supposed to
do so.

Test: atest
WmTests: com.android.server.wm.DisplayContentTests#testRecentsNotRotatingWithFixedRotation
Fixes: 157447718
Change-Id: I3f2a46ddfe5e76ad3200ea27ad272ae36b6bcc8d

Former-commit-id: a8849f27bc9f85a5cb441ab155ae0afe7c99872a