Animate screenshot dismiss button appearance
Currently the dismiss button just pops in once the animation
is complete, which is somewhat jarring (and because the animation
decelerates as it ends, it feels like there's a delay before the
button actually appears).

This change starts fading in the button while the animation is
happening, matching its location to the preview at each animation
update.

Bug: 155415756
Fix: 155415756
Test: manual -- tested with ltr/rtl, portrait/landscape, and from
overview vs keychord, all looked correct

Change-Id: Ibd3ea2a4127d6fee87ec9b73d73f8f5a184d1417

Former-commit-id: 958eb9dcd5c3096e3ce6e109e97682d74c835a52