Allow the active IME to see the target app package
This CL ensures that API calls from IMEs will not be blocked by the
package visibility restriction introduced in Android 11, as long as
the IME is or was interacting with the target app.

This types of special allow rules are called "implicit visibility"
internally, because explicit <queries> elements in
ApplicationManifest.xml aren't required in such cases, and this CL
adds one more case to this "implicit visibility" scenario so that IMEs
can have "implicit visibility" to the target apps.

Fix: 152909969
Test: atest CtsInputMethodTestCases:PackageVisibilityTest
Test: atest CtsInputMethodTestCases:PackageVisibilityTest --instant
Change-Id: I28ccdd7409d167cb6d277135e9ebaed5b2422d79

Former-commit-id: 724c5bb8afaf4ea81f421c854117c30107225b96