Don't throw exceptions in NotifCollection when initializing
When systemUI is in the process of connecting to system server to
receive notifications, there's a short window in which we might receive
events out of order. These events will violate our otherwise strict
requirements. To solve this, we really need a more robust way to connect
to system server, but in the meantime, we create a "forgiveness window"
that lasts five seconds. During this time, things that would normally
crash the system only log an error.

Bug: 151317347
Test: atest, manual
Change-Id: I9aad4ec94826189efc13433b4cd38f443113fd7e

Former-commit-id: 203f80a0b2d311328baef4f94f44dd07064738a1