Don't clear mUsingBLASTSyncTransaction until sending callback
If a window has drawn, it doesn't mean the client will immediately get
the callback. The applySync transaction may be waiting on other windows
to draw. In that case, we want to continue adding any transaction to the
blast sync transaction to ensure we don't submit changes until the
client gets a chance to handle the transaction. This is done by delaying
when mUsingBLASTSyncTransaction is set to false so all requests for the
WindowContainers are still added to the syncTransaction

Additionally, don't report back that the window completed the draw until
we allow for one more performSurfacePlacmement. It's possible that the
window has other changes that need to be added to the syncTransaction so
only notify onTransactionReady when we allow one more pass to gather the
transaction information.

Test: Split screen
Test: WindowOrganizerTests
Fixes: 158793236
Change-Id: If0a4071869fa1ba70f59aad25606a6e3cb7dafee

Former-commit-id: 5452169c16889d6ab11aa1b02d26df1c49350458