Use actual resizability for split-screen operations
Split-screen and activity-start were relying on resizeMode,
but that is what the activity requests vs how the system
actually treats a given task/activity. This caused some
inconsistencies in how non-resizable things were handled
when the developer option to force resizability was turned
on.

Also needed to fix an issue where non-leaf tasks would
get stuck with an out-dated intent. In particular, this
happened with the root HOME task which meant its
resizeMode could never change from that of FallbackHome.

Bug: 158514252
Test: enable force resizable in dev options and then launch
      camera into secondary split.
Change-Id: If06d047813315f39894bead19d2cbe0c9444f254

Former-commit-id: 889f49683f02a5285d4a47a7f1686368697527e8