Reevaluate some carrier-associated apps' status on SDK changes.
If a carrier-associated app is added to the system image after the
device is initially launched (e.g. Q -> R OTA), the logic to disable
apps without a corresponding SIM will not run, leaving the app in an
enabled state in more cases than there should be.

To safely account for this, we only reevaluate on SDK change for apps
whose addedInSdk value falls within the range since the last time we
evaluated carrier apps' statuses. Apps that set lower (or higher)
SDK versions are ignored. We choose *not* to use minimum or target SDK
version here because it's more likely to change with each OTA /
prebuilt, which could cause unintended disable operations (or lack
thereof).

Bug: 154872019
Test: manual, QA, atest FrameworksTelephonyTests:CarrierAppUtilsTest
Change-Id: I97ddfa03f34e7d00e66ce7cd508d6d2a5e83fbc2

Former-commit-id: 87cd4b076176539d5e3520e4a10dfaa380380c52