Properly collapse QQS header when QQS is disabled
In landscape, we need to fully hide the QQS when the user starts to do
an inline reply, otherwise nothing fits on the screen. This CL adds back
in some of the old Q behavior wherein we set the height of the header to
essentially zero when the QQS has been disabled (in this case, by the
inline reply logic).

Fixes: 157150870
Test: manual
Change-Id: Ib14603eb2e67bf4bf3b624b677a74754d4e54f41

Former-commit-id: 59a62fc264f249167607ca2475f35ed019b0fa25