Fixed the conflict of the release methods of ActivityView & TaskEmbedder
There is a case in Bubble that we might release a un-initialized
TaskEmbedder while calling ActivityView.release() (ex. create a bubble
and don't expand it and then dismiss it) which will cause an exception.

Allow ActivityView.release() to be called at anytime and check the
initialization state before calling TaskEmbedder.release().

Bug: 155417004
Test: 1. Add bubbles and don't expands them then dimiss them
      2. Check logs.
Change-Id: I583ddd36f407c908cb63c109834940fc91c8d130

Former-commit-id: 2806269a5a4bb08c1f8694a4b601256db90728b1