Change BroadcastDispatcher to 1receiver/action
With the aggregated filter, as it had to be registered when any single
action changed (added for first time or removed the last receiver for
that action), if there was any action that got periodically
registered/unregistered, it would cause frequent cycles of
unregister/register with Context. This leads to interruption of service.

With this change, changes in one action would not affect receivers for
the others.

Test: manual, SystemUI works as expected
Test: manual, dumps and logs
Test: atest UserBroadcastDispatcherTest ActionReceiverTest
Bug: 148390204
Bug: 157165818

Change-Id: I7961f6c60600dfa0407615319c43bb72e43b4795

Former-commit-id: f3fb89347aa5631f595a841d05856c17856b4674