Fix issue with dragging bubbles in fullscreen.
This ensures that we only animate the stack temporarily invisible if we're not dragging something or a flyout is still showing.

This also adds a 2s delay before hiding the stack if it was visible due to drag/flyout, since a) it looks better b) that way you can open the stack back up if you didn't mean to let go.

Fixes: 159064066
Test: open fullscreen youtube, show status bar, drag around until status bar goes away, then fling bubbles somewhere and observe
Test: get a flyout while in fullscreen, touch the flyout until the status bar goes away, then collapse flyout
Change-Id: I06bc148161a1d51ee047b3861c1d01e7685c2943

Former-commit-id: 44318ebe363aa968d81994d6b13073bf4c0ce982