Repopulate cache with sibling visibility on remove
This change ensures that we re-add the visibility state of a shared user
id when one of its members is removed. Prior to this change, we were
removing the app ID (and thus all other members of the shared uid) but
only recomputing on the removed setting, which we ignore in the update
(essentially an expensive no-op). This change ensures that we run
through the shared settings and re-compute with them instead.

Fixes: 159197157
Test: atest AppEnumerationTests SharedUidPermissionsTest
Change-Id: I85250e4264888ac8583a2b4798ff6a2f93d6a74d
(cherry picked from commit a6dbc1b150ad47f03c1186737422d146f290dc7c [formerly 5c5ac45ea00d469954752c276fa32e7fd0cc1b63])

Former-commit-id: c6985445ad7cc3c355102f5023d574e1346f0bcf