Allow staging multi-package session immediately after abandonment
When we stage a session, we compare it with other already staged
sessions to detect overlap. We do not allow a session to stage if it
contains a package which is already staged in another session.

Due to a bug, while comparing we were not excluding sessions whose
parents were abandoned already. This CL fixes that.

Bug: 159409632
Test: atest
StagedInstallTest#testStageAnotherSessionImmediatelyAfterAbandonMultiPackage

Change-Id: I740986af666b3d2d8ab832c4bcdb855c1ed3c16f

Former-commit-id: 806ccabb73e93ad08a5284a7d651d6827aefc6ac