Add support for floating IME to split-screen
This basically converts a floating IME into what looks like
an ime that has some small size below the bottom of the screen
so that it animates to 0.

Also added alpha for this case. Would add it for all cases,
but that doesn't work well for split-screen since it translates
windows rather than insetting them so you see black flickers.

Bug: 159459047
Test: Use an expanded-ui ime in landscape and see a shortened
      animation + fade-in.
Change-Id: Ib1da2750689a4bd4d3a3d4e9bc1e29aa8d41376c

Former-commit-id: 437a879af3e34e38fb17dcedd4db6b6b98d57dfc