Remove duplicate coordinate rotation code for screenshots.
The assist screenshot code path request the screenshot from
DisplayContent. DC.screenshotDisplayLocked was rotating the coordinates
to compenstate for the screenshot always taken in default orientation.
However, SurfaceControl.screenshot already takes care of that so we
ended up rotating the frame twice and flipping 180 and 90 twice. This
caused the screenshot to get taken with a crop for portrait and the
orientation flipped.

Test: Assist screenshot is correct now
Fixes: 159229894
Change-Id: Ie24eca4dd62225d106dbeafd96afba15de9fd623

Former-commit-id: 161eaff04c8f82c453b28b4439c1d865c9498a57