Stop sharing ProximitySensor in Doze code.
One part of the Doze code would pause the sensor, such that when
another tried to use it, it would never get a result back.

The DozeTriggers tried to use the sensor to determine if it should
wake the screen up when a notification comes in. Because it never
got a result, it never woke the screen up, meaning that the
notification would never be visible.

Fixes: 156874080
Test: atest SystemUITests && manual.
Change-Id: I0c5927f102237c57e9e50bc46ba7ebcb7fb79cec

Former-commit-id: 2fb102a94b3ecff7bad849e21802a4b5e3c3c50a