Making transitions between gone hosts work better
The transition previously wasn't aligned with the
content at all and the landscape transition was
broken as the content appeared out of no-where.
The main reason was that we were still changing
the bounds of the player to the invisible player
when not in a guided transition.

This also fixes issues where we were not caching
aggressively enough, because only a subset of the
state is actually relevant for the measurement

Bug: 159834486
Test: have resumable player, drag down both in landscape and portrait
Change-Id: I1671c9d48e7f1b1166e4505e90785bc699528d1c

Former-commit-id: 85d1f8bfee61d2da14ac60766f22e8a57040a2a9