UsbDebuggingActivity: notify failure in onDestroy, not onStop.
On initial boot, the prompt can be shown and then get hidden, which
would previously cause the initial public key authorization request to
be spuriously rejected, if the adb client was running during boot. Move
the rejection to onDestroy instead, so we only reject prompts that have
no chance of being actually accepted.

Bug: http://b/159061108
Test: manual
Change-Id: I9acb1145b1f8a6c8d6f82d1c4459bcd86eb306df

Former-commit-id: 7b66783514ff6eecbb7033d4d318f2255983cea1