Remove mEnterAnimationPending check for accessibility
Accessibility was relying on WS.mEnterAnimationPending to determine if a
window was ready on screen. However, that variable no longer represents
what it used to and it doesn't actually mean the app is animating. The
variable is only set when a relayout is requested and then unset when
show is called. Accessibility is already checking isOnScreen and
isVisibleLw so that should cover the cases to check if the window is
ready.

mEnterAnimationPending is not always set properly and causes issues with
accessibility, which is the only place currently using it.

Fixes: 152537982
Test: Magnification for Accessibility works
Change-Id: Ib0f37b3cc6d744fd84f466a65e6ce006e07e116e

Former-commit-id: 4bf251ca4ea094052ecb88ad053dc678655f8903