Delay sending accessibility events a little for temporarily detached view
In RV, a view could be temporarily detached in one draw cycle.
Accessibility events sent during this time will be lost. To prevent the
events from being lost, we will set a flag when a view is detached by
parent. And when sending accessibility events for detached view, we will
delay it a little by
ViewConfiguration.getSendRecurringAccessibilityEventsInterval. This way,
if the view is attached in one draw cycle, the events will be sent out.

Also add more setStateDescription() in CompoundButton to for safer code.
(this is not the cause of this bug though).

Fix: 151125936

Test: tested that the bug is fixed.

Change-Id: Iffca8c87bad4fa2f66862b966e351562d77d6d76

Former-commit-id: e34bc33af3a7e1ca75037441fb00f0662a9f2886