Migrate initial set of flags to DeviceConfig.
This migrates the flags defined inside JobSchedulerService to
DeviceConfig. DeviceConfig is the recommended way and using it means
that we only operate/parse the set of flags that changed instead of
every single flag whenever one of them changes.

The flags inside QuotaController can be moved in here since only
individual flags are operated on now and we can do everything with
just one listener.

Bug: 124466289
Test: atest CtsJobSchedulerTestCases
Test: atest FrameworksMockingServicesTests:JobSchedulerServiceTest
Test: atest FrameworksServicesTests:MaxJobCountsTest
Test: atest SettingsProviderTest:SettingsBackupTest
Change-Id: I16b937df6a8eae338cfd77e72bcd4d264036db82

Former-commit-id: 1c086b48a03c3569fe4b207b9b33489a47715c85