Fix stack reference is not updated when task organizer is dead.
The Task Organizer controller handles the binder dead when SystemUI
crashed. mTaskOrganizer was assigned to null first and then removes
the related child. But the stack reference is not updated if
taskOrgainzer is null. The invalid home task is referenced when we
start the launcher.

This CL updates the condition to make sure the stack references are
updated.

Bug: 158756447
Test: 1. Setup the split-screen.
       2. Press power key to show Lock-screen
       3. Kill the systemUI process
       4. Unlock and resume to the launcher.
Change-Id: I5c3f03c4d7fecab6e812dfba6823d79d844d6a26

Former-commit-id: 1cfcdf7bc963e8a34cac1d5b46fcbf78cb5108b8