Fixing an embedded display as a top focused display
There's no window observer for an embedded display. All windows of an
embedded display would be sent to the window observer of its parent
display. So when the focused window is at an embedded display, the
focused display should be its parent display, not the embedded display.
Otherwise the active window and top focused window could not be set
if the display of the received windows isn't the top focused display.
Besides adding more information for debugging when dumping the
accessibility info.

Bug: 160388022
Test: a11y CTS & unit tests
Test: manual testing for bubble expanded with talkback on
Change-Id: I1de53e4132fc935e3f188271c8a716fefc28294a

Former-commit-id: edbfd229e83280ede42fcd35e9db0aab5f2a88a1