Offsets the position of window maganification by the margin
When window magnification is toggled on, the position of window
magnification is not aligned with the one after updating its
layout params. And this bug looks like a transition animation that
is unexpected.
To fix it, we have to offset the position by the margin first
when adding window magnification layout before updating its
layout params.

Bug: 160939052
Test: none
Change-Id: I8e83094d39f33b670e160aaf11c3d09d60dc77bf

Former-commit-id: 0ede1eadb77f469a7fd7843fcca45c3c144cc004