Use reentry size to calculate PIP entry bounds
Use the saved reentry size when restoring the PIP
bounds if available.

Note: this CL doesn't actually change the size of
the bounds we restore to because the saved reentry
size is always the normal bounds. It does however
make it so that if the saved reentry size were to
be changed to something other than the normal bounds
then it would actually restore to it on reentry.

Bug: 160799929
Test: atest SystemUITests
Test: in an environment where the saved reentry size
is changed, the PIP window is restored to that size
when re-entering PIP

Change-Id: I0bbb4725c249d9d81b80b46d1bff5d6cd11981c3

Former-commit-id: 4177a5ee62fe548d3c90963eb8b78e29f7bda707