Use VariationPosition instead of Axis.
Skia's SkFontArguments started out with an 'Axis' type to provide
variation position. This is an unfortunate name as the values are
actually positions on the axis (do not describe the axis itself) and the
collection of them is commonly known as the variation. Skia would like
to remove this long deprecated type, so move to the new interface.

Change-Id: I460ca8cb938f078d25f41b4cd1dfc4db3a4e824f

Former-commit-id: 78d58fd0347dcce8556a982f54c7069020d0bcdc