Prevent NotificationIconContainer$3 from retaining NotificationEntry
It has been observed in heap dumps that NotificationIconContainer$3 can
retain a NotificationEntry after the entry has been dismissed. To
prevent this, this change breaks the reference chain between
NotificationIconConatiner$3 and the IconState for the status bar icon.
The cut is performed by setting the animation end action to null after
the animation is started.

Fixes: 160588607
Bug: 161119435
Test: manual - Clear all notifications. Take heap dump. Verify that
weakref in AnimationEndAction has been collected (ie. is null).

Change-Id: I32184a30df89a5172876a43da4896a2e06e3c86c

Former-commit-id: a455113a2b5e3d965c134312d59b1d9d4d59f84f