Fix UnsupportedOperationException in CutoutSpecificationBenchmark
Exception was due to an exception from trying to get the display from
the app context, instead of an Activity context.

java.lang.UnsupportedOperationException: Tried to obtain display from a Context not associated
 with one.

In order to fix that and fix the display metrics factor, the CutoutSpecificationBenchmark only run
for the resolution 1080x1920 and density is 3.5f.

Test: atest CutoutSpecificationBenchmark

Change-Id: I0be9d09a859c5b3a8c5abff11a3d06d809edf9ab

Former-commit-id: b594b6f09b582722221a2b269dfb62ed5bb723a7