Remove service stub dep on default libs
Not specifying any sdk_version defaults the library to building
against the platform internals. We do not want to build the stubs
against the platform internals, as that risks leaking internals
in the stubs.

Follow what the other stubs are doing, which is to build against
no default libs, and compiling with java version 1.8.

Bug: 158850553
Test: m
Change-Id: I310842747422b108ac2a1fcf11afac5c323ed5b1

Former-commit-id: 4304e6065526c28581ddde18481146eecb12b531